requirements.txt
Django==2.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0

